added component test and renamed file
